Address draft PR validation
